Applied fixes for PR reviews. Test updated.
